requirements.txt
playwright==1.49.0
supabase==2.10.0
transformers>=4.45.0,<4.47
torch>=2.0.0
pillow>=10.0.0
python-dotenv==1.0.1
httpx>=0.27.0
